Add modal for stake tx history
